seperating or deleting unused functions, adding logging config, adding optional imports
