+++ scripts/alternate_graph_methods.py
@@ -2,6 +2,12 @@ from __future__ import annotations
 
 
 class AlternativeCylCollection:
+    """
+    This class contains alternative implementations 
+        of various functions in the CylinderCollection class.
+    This class is used for memory and computational 
+        efficiency efinement
+    """
     def initialize_graph(self):
         """This function initializes edge attributes as cylinder objects."""
         gr = nx.Graph()
