+++ src/canhydro/CylinderCollection.py
@@ -2,36 +2,36 @@ from __future__ import annotations
 
 import copy
 from itertools import chain
+
 import networkx as nx
 import numpy as np
-
 from shapely.geometry import Point
 from shapely.ops import unary_union
 
 from src.canhydro.Cylinder import create_cyl
 from src.canhydro.DataClasses import Flow
-from src.canhydro.geometry import (concave_hull, draw_cyls, furthest_point)
+from src.canhydro.geometry import (concave_hull, draw_cyls, 
+                                    furthest_point,
+                                    get_projected_overlap)
 from src.canhydro.global_vars import config_vars, log
 from src.canhydro.utils import intermitent_log, lam_filter, save_file
-
-from src.canhydro.global_vars import config_vars, log
-from src.canhydro.utils import intermitent_log, lam_filter, save_file
-
 from src.canhydro.import_options import _try_import
 
-#Optional imports 
-has_pyplot = _try_import('pyplot', 'matplotlib')                                   
-has_mem_prof = _try_import('LogFile', 'memory_profiler')
-has_GeoSeries = _try_import('GeoSeries', 'geopandas')
-has_scipy_dist_proj = _try_import('distance', 'scipy.spatial')
+#Optional imports    
 
-has_proj_overlap = _try_import('get_projected_overlap', 'src.canhydro.geometry')
-has_pool_proj = _try_import('pool_get_projection', 'src.canhydro.geometry')
-has_vectorized_proj = _try_import('vectorized_get_projection', 'src.canhydro.geometry')
+if has_geopandas := _try_import('geopandas'):
+    from geopandas import GeoSeries
 
-if has_mem_prof:
+if has_mem_profiler := _try_import('memory_profiler'):
+    from memory_profiler import LogFile
     import sys
     sys.stdout = LogFile()
+
+if has_matplotlib := _try_import('matplotlib'):
+    import matplotlib.pyplot as plt
+
+if has_spatial := _try_import('scipy.spatial'):
+    from scipy.spatial import distance
 
 NAME = "CylinderCollection"
 
@@ -43,7 +43,6 @@ class CylinderCollection:
     # initialize our object level variables for cylinder objects
     def __init__(self) -> None:
         self.file = ""
-        # self.collection = CylinderList()
         # Aggregate values from file
         self.surface_area = np.nan
         self.file_name = ""
@@ -51,14 +50,11 @@ class CylinderCollection:
         self.avg_sa_to_vol = np.nan
         self.max_branch_order = np.nan
         self.max_reverse_branch_order = np.nan
-        self.canopy_scope = np.nan  # desc of canopy
         self.extent = {
             "min": [np.nan, np.nan, np.nan],
             "max": [np.nan, np.nan, np.nan],
         }
         # to populate with x,y,z mins and maxs
-        self.aggregate_angle = np.nan
-        self.descriptive_vectors = np.nan  # Average, median, mode vectors
         self.treeQualities = {
             "total_psa": None,
             "tot_hull_area": None,
@@ -72,7 +68,6 @@ class CylinderCollection:
 
         # Projection Attrs
         self.projections = {"XY": False, "XZ": False, "YZ": False}
-        self.stem_path_lengths = []
         self.hull = None
         self.stem_hull = None
 
@@ -83,14 +78,14 @@ class CylinderCollection:
         # Graph and Attributes
         self.graph = None
         self.digraph = None
-        self.min_graph = None
         self.flows = None
         self.divide_nodes = None
         self.drip_nodes = None
         self.cyl_to_drip = None
         self.drip_points = {"x": np.nan, "y": np.nan, "z": np.nan, "flow_id": np.nan}
         self.flow_to_drip = {
-            0: 1
+            0: 1 #allows us to bootstrap stemflow 
+                    # (cyl 1 drains to 0, so its child cyl 2 drains to 0, etc)
         }  # A dictionary of flow ids with values equal to their drip node ids
         self.trunk_nodes = []
         self.drip_point_loc = None
@@ -99,15 +94,7 @@ class CylinderCollection:
         self.drip_flow_components = None
         # Calculations using graph results
         self.flow_chars = {}
-        self.divide_points = []
-        self.stemPolys = []
-        self.compGraphs = []
         self.trunk_lean = None
-
-    def aggregate_characteristics(self):
-        """Calculates the summations, averages etc. of cylinder characterictics
-        that might be of interest"""
-        return True
 
     def from_csv(self, file, aggregate_cyls=True):
         """Initializes a new Cyl Collection based on the data in a QSM
@@ -141,8 +128,6 @@ class CylinderCollection:
                 "min": [min_x, min_y, min_z],
                 "max": [max_x, max_y, max_z],
             }
-
-        self.descriptive_vectors = np.nan  # Average, median, mode vectors
 
         self.theta = np.nan
         log.info(f"{file.name} initialized with {self.no_cylinders} cylinders")
@@ -295,16 +280,6 @@ class CylinderCollection:
         else:
             self.hull = hull
 
-    def initialize_graph_from(self):
-        """This function creates an undirected_graph and initializes edge attributes as cylinder objects"""
-        gr = nx.Graph()
-        edges = (
-            (int(cyl.cyl_id), int(cyl.parent_id), {"cylinder": cyl})
-            for cyl in self.cylinders
-        )
-        gr.add_edges_from(edges)
-        self.graph = gr
-
     def initialize_digraph_from(
         self, in_flow_grade_lim=config_vars["in_flow_grade_lim"]
     ):
@@ -348,12 +323,12 @@ class CylinderCollection:
         """
         trunk_nodes = self.get_trunk_nodes()
 
-        trunk_contraction, titans = self.contracted_nodes(trunk_nodes, 0, self.graph)
+        trunk_contraction, _ = self.contracted_nodes(trunk_nodes, 0, self.graph)
         trunk_paths = nx.shortest_path(trunk_contraction, target=0)
         dists = {node: len(path) - 1 for node, path in trunk_paths.items()}
         return dists
 
-    def find_flow_components(self, inFlowGradeLim=-1 / 6):
+    def find_flow_components(self):
         g = self.digraph
         if type(g) == nx.Graph:
             msg = "Find Flow Digraph invoked for undirected graph"
@@ -411,10 +386,8 @@ class CylinderCollection:
                 attr["cylinder"].cyl_id
                 for _, _, attr in component_graph.edges(data=True)
             ]
-            for id in component_cyls:
-                # if cyl_to_drip_node[id]:
-                # log.warning(f"Cylinder with id {id} has two or more identified to drip nodes: {cyl_to_drip_node[id]} and {drip_node}")
-                cyl_to_drip_node[id].append(drip_node)
+            for idx in component_cyls:
+                cyl_to_drip_node[idx].append(drip_node)
 
             drip_components.append((drip_node, g_drip.subgraph(component_nodes).copy()))
 
@@ -682,28 +655,10 @@ class CylinderCollection:
 
         return statistics
 
-    def describe(self, metric: str, a_lambda: function = lambda: True, **args) -> dict:
-        """
-        Takes in a metric of the tree and a filter and returns summary stats
-        for the given input metric
-        Potentially outputs a histogram if requested?
-        """
-        return {True}
-
-    def compare(
-        metric: str,
-        a_lambda: function = lambda: True,
-        b_lambda: function = lambda: True,
-    ):
-        """
-        Compares the two lists of metrics defined by the two input functions
-        """
-
     def get_drip_points(
         self,
         # metric :str = 'projected_area',
-        percentile: int = 30,
-        **args,
+        percentile: int = 30
     ):
         """
         Returns the locations of the drip points.
@@ -751,70 +706,74 @@ class CylinderCollection:
         a_lambda: function to filter drip points displayed (e.g. those with projected area>10m^2 )
         scale: how large of a boundary to draw around drip points
         """
-        # drip_point_locs = self.get_drip_points()
-        # drip_point_locs_x = [pt[0] * scale for pt in drip_point_locs]
-        # drip_point_locs_y = [pt[1] * scale for pt in drip_point_locs]
-        # drip_point_locs_xy = [[pt[0] * scale, pt[1] * scale] for pt in drip_point_locs]
+        drip_point_locs = self.get_drip_points()
+        drip_point_locs_x = [pt[0] * scale for pt in drip_point_locs]
+        drip_point_locs_y = [pt[1] * scale for pt in drip_point_locs]
+        drip_point_locs_xy = [[pt[0] * scale, pt[1] * scale] for pt in drip_point_locs]
 
-        # math.floor(np.min(drip_point_locs_x))
+        math.floor(np.min(drip_point_locs_x))
 
-        # mins = self.extent["min"]
-        # maxs = self.extent["max"]
-        # extents = [mins[0], maxs[0], mins[1], maxs[1]]
-        # # min_xy = np.min(mins)
-        # # max_xy = np.max(maxs)
-        # # x_mesh, y_mesh = np.meshgrid(
-        # #     np.arange(min_xy, max_xy, 0.05), np.arange(min_xy, max_xy, 0.05)
-        # # )
-
-        # min_xy = np.min(
-        #     [
-        #         math.floor(np.min(drip_point_locs_x)),
-        #         math.floor(np.min(drip_point_locs_y)),
-        #     ]
-        # )
-        # max_xy = np.max(
-        #     [math.ceil(np.max(drip_point_locs_x)), math.ceil(np.max(drip_point_locs_y))]
-        # )
+        mins = self.extent["min"]
+        maxs = self.extent["max"]
+        extents = [mins[0], maxs[0], mins[1], maxs[1]]
+        # min_xy = np.min(mins)
+        # max_xy = np.max(maxs)
         # x_mesh, y_mesh = np.meshgrid(
-        #     np.arange(min_xy, max_xy, 0.005), np.arange(min_xy, max_xy, 0.005)
+        #     np.arange(min_xy, max_xy, 0.05), np.arange(min_xy, max_xy, 0.05)
         # )
 
-        # def dist_to_drip(a, b):
-        #     distances = distance.cdist([[a, b]], drip_point_locs_xy)
-        #     min_dist = np.min(distances)
-        #     return math.log(1 / min_dist)
+        min_xy = np.min(
+            [
+                math.floor(np.min(drip_point_locs_x)),
+                math.floor(np.min(drip_point_locs_y)),
+            ]
+        )
+        max_xy = np.max(
+            [math.ceil(np.max(drip_point_locs_x)), math.ceil(np.max(drip_point_locs_y))]
+        )
+        x_mesh, y_mesh = np.meshgrid(
+            np.arange(min_xy, max_xy, 0.005), np.arange(min_xy, max_xy, 0.005)
+        )
 
-        # distance_matrix = np.zeros((x_mesh.shape[0], x_mesh.shape[0]))
+        def dist_to_drip(a, b):
+            distances = distance.cdist([[a, b]], drip_point_locs_xy)
+            min_dist = np.min(distances)
+            return math.log(1 / min_dist)
 
-        # for a in range(x_mesh.shape[0]):
-        #     for b in range(x_mesh.shape[0]):
-        #         distance_matrix[a][b] = dist_to_drip(x_mesh[b][a], y_mesh[b][a])
+        distance_matrix = np.zeros((x_mesh.shape[0], x_mesh.shape[0]))
 
-        # fig, ax = plt.subplots()
+        for a in range(x_mesh.shape[0]):
+            for b in range(x_mesh.shape[0]):
+                distance_matrix[a][b] = dist_to_drip(x_mesh[b][a], y_mesh[b][a])
+        
+        if has_matplotlib:
+            _, ax = plt.subplots()
 
-        # ax.contourf(
-        #     y_mesh,
-        #     x_mesh,
-        #     distance_matrix,
-        #     levels=15,
-        #     max=0.5,
-        #     cmap=plt.cm.Blues,
-        #     extend="neither",
-        #     extent=extents,
-        # )
+            ax.contourf(
+                y_mesh,
+                x_mesh,
+                distance_matrix,
+                levels=15,
+                max=0.5,
+                cmap=plt.cm.Blues,
+                extend="neither",
+                extent=extents,
+            )
 
-        # ax.scatter(drip_point_locs_x, drip_point_locs_y)
+            ax.scatter(drip_point_locs_x, drip_point_locs_y)
 
-        # filtered_cyls, _ = lam_filter(self.cylinders, a_lambda, return_all=False)
-        # polys = [cyl.projected_data["XY"]["polygon"] for cyl in filtered_cyls]
-        # # breakpoint()
-        # if len(polys) > 0:
-        #     geoPolys = GeoSeries(polys)
-        #     geoPolys.plot(ax=ax)
-        # else:
-        #     log.warning(
-        #         "Drip Map: No cylinders returned for lambda function: {a_lambda}"
-        #     )
-        # plt.show()
-        print('Drip map called')
+            filtered_cyls, _ = lam_filter(self.cylinders, a_lambda, return_all=False)
+            polys = [cyl.projected_data["XY"]["polygon"] for cyl in filtered_cyls]
+            
+            if len(polys) > 0:
+                geoPolys = GeoSeries(polys)
+                geoPolys.plot(ax=ax)
+            else:
+                log.warning(
+                    "Drip Map: No cylinders returned for lambda function: {a_lambda}"
+                )
+        else:
+            log.warning(
+                "Unable to draw drip map as module matplotlib.pyplot is not available"
+            )
+        return distance_matrix
