+++ src/canhydro/utils.py
@@ -5,44 +5,60 @@ import os
 import shutil
 import stat
 
+
 import numpy as np
-# from numba import njit, prange
-# from numba.typed import List
 
 from src.canhydro.global_vars import input_dir, log, output_dir, time_stamp
+from src.canhydro.import_options import _try_import
+
+ 
+has_numba = _try_import('numba')
+if has_numba:
+    from numba import njit, prange
+    from numba.typed import List
 
 
-# def stack(to_stack:list[np.array], col: bool == True):
-#     """
-#         A wrapper for njit stack that handles errors and allows for 
-#         less strict typing 
-#     """
-#     list_of_array = List(to_stack)
-#     try:
-#         njit_stack(list_of_array,col)
-#     except ValueError as err:
-#         left_shape = list_of_array[0].shape[0]
-#         right_shape = list_of_array[1].shape[0]
-#         stack_type ='column' if col else 'row'
-#         msg = f'{err}: Cannot {stack_type} stack arrays with shapes {left_shape} and {right_shape}'
-#         log.error(msg)
-#         raise ValueError(msg) from err
-
-# @njit()
-# def njit_stack(list_of_array:np.array[np.array()], col: bool):
-#     """
-#     numba doesn't play well with np stacks, so I had to do it myself
-#     """
-#     num_in = len(list_of_array)
-#     left_shape = list_of_array[0].shape[0]
-#     shape = (num_in, left_shape)
-#     stacked_array = np.empty(shape)
-#     for j in prange(len(list_of_array)): 
-#         stacked_array[j] = list_of_array[j]
-#     return stacked_array if not col else stacked_array.T
+def stack(to_stack:list[np.array], col: bool = True):
+    """
+        A wrapper for njit stack that handles errors and allows for 
+        less strict typing 
+    """
+    if not has_numba:
+        non_njit_stack(to_stack, col)
+    else:
+        list_of_array = List(to_stack)
+        try:
+            njit_stack(list_of_array,col)
+        except ValueError as err:
+            non_njit_stack(to_stack, col)
 
 
-def on_rm_error(func, path, exc_info):
+def non_njit_stack(to_stack:list[np.array], col: bool = True):
+    list_of_array = list(to_stack)
+    num_in = len(list_of_array)
+    left_shape = list_of_array[0].shape[0]
+    shape = (num_in, left_shape)
+    stacked_array = np.empty(shape)
+    for idx, _ in enumerate(list_of_array): 
+        stacked_array[idx] = list_of_array[idx]
+    return stacked_array if not col else stacked_array.T
+
+if has_numba:
+    @njit()
+    def njit_stack(list_of_array:np.array[np.array()], col: bool):
+        """
+        numba doesn't play well with np stacks, so I had to do it myself
+        """
+        num_in = len(list_of_array)
+        left_shape = list_of_array[0].shape[0]
+        shape = (num_in, left_shape)
+        stacked_array = np.empty(shape)
+        for j in prange(len(list_of_array)): 
+            stacked_array[j] = list_of_array[j]
+        return stacked_array if not col else stacked_array.T
+
+
+def on_rm_error(path):
     # path contains the path of the file that couldn't be removed
     # let's just assume that it's read-only and unlink it.
     os.chmod(path, stat.S_IWRITE)
@@ -52,9 +68,6 @@ def on_rm_error(func, path, exc_info):
 def create_dir_and_file(filename,) -> None:
     print(type(filename))
     os.makedirs(filename, exist_ok=True)
-    # f = open(filename, "w+")
-    # f.write("Now the file has more content!")
-    # f.close()
 
 
 def del_dir(filename) -> None:
@@ -65,8 +78,7 @@ def read_file_names(file_path=input_dir):
     """Reads in filenames to list"""
     paths = sorted(file_path.iterdir(), key=os.path.getmtime)
     fileNames = [f.name for f in paths if f.suffix == ".csv"]
-    print(paths)
-    print(fileNames)
+    log.info(f'fileNames found in {file_path} on read: {fileNames}')
     return fileNames
 
 
@@ -81,11 +93,16 @@ def save_file(
     dir = "/".join([str(output_dir), method, ""])
     ofname = "_".join([file, method])
     ofname_ext = "".join([ofname, fileFormat])
-    aggname = "_".join(["agg", method, fileFormat])
-    aggname_ext = "".join([aggname, fileFormat])
     folderExists = os.path.exists(dir)
     fileExists = os.path.exists(dir + ofname_ext)
-    aggExists = os.path.exists(dir + aggname_ext)
+
+    #Leaving the below in as a reminder 
+    # to add 'append to aggregate file' 
+    # functionality back in 
+    # aggname = "_".join(["agg", method, fileFormat])
+    # aggname_ext = "".join([aggname, fileFormat])
+    # aggExists = os.path.exists(dir + aggname_ext)#
+
     if not folderExists:
         os.makedirs(dir)
 
@@ -130,8 +147,17 @@ def intermitent_log(prog: int, whole: int, msg: str, freq: int = 0.0001):
 
 
 def lam_filter(objects, a_lambda: function, return_all: bool = False):
-    """Takes in a lambda that filters on cylinder attrs
-        returns a list of cylinders meeting that requirements 
+    """
+        Takes in a lambda that filters on cylinder attrs
+        returns a list of cylinders for which that
+        lambda func returns true
+        i.e. lam_filter(cylinders, lambda: diameter > 0.5)
+                -returns all cylinders with diameter > 0.5
+
+        return_all: bool 
+            - if true, returns all cylinders along with 
+                a boolean array of the same length w/ the 
+                results of the lambda function passed
     """
     if a_lambda.__code__.co_name != "<lambda>":
         raise ValueError("a lambda required")
